Create codeql-analysis.yml (#648)
This PR enables CodeQL, a semantic code analysis engine by Github, to scan for security issues on this repo.

By default it will scan on the following interval:

- Every new commit against the default branch (in this case master)
- Every new pull request against the default branch (in this case master)
- Once per week on **Sunday at Midnight UTC** which can be seen in the newly created .github/workflows/.codeql-analysis.yml file under the -cron flag under schedule. 
	
To learn more about the analysis tool and why we are using it please see: https://codeql.github.com/docs/codeql-overview/ for information about codeQl, and [this](https://github.blog/2020-05-06-new-from-satellite-2020-github-codespaces-github-discussions-securing-code-in-private-repositories-and-more/) blog post for a little more context and background. This tool has already proven useful in the repos in which it has been enabled and has found bugs which we previously had missed.