Add timestamps to the nitro-shim log output
After launching the enclave application `run.sh` execs `sleep.sh`
which periodically polls the enclave state and exits if it dies,
propagating the failure to the container.

In debugging, it's difficult to match up events in the container
log, which are largely the trace output of the run and sleep
scripts, so add periodic timestamps to it's clear how long things
were running.